resolve remaining merge git diff --name-only --diff-filter=U
