+++ src/defaults/syncSettings/syncTimeScaleSettings.js
@@ -2,9 +2,10 @@ import { format, time } from 'd3';
 
 export default function syncTimeScaleSettings(settings) {
     //Coerce invalid time scale arguments.
-    settings.time_scale = ['date', 'day'].indexOf(settings.time_scale.toLowerCase()) > -1
-        ? settings.time_scale.toLowerCase()
-        : 'date';
+    settings.time_scale =
+        ['date', 'day'].indexOf(settings.time_scale.toLowerCase()) > -1
+            ? settings.time_scale.toLowerCase()
+            : 'date';
 
     //Define settings variables to handle both date and day time scales.
     if (settings.time_scale === 'date') {
